Moved influxdb-pojomapper classes to parent package #340
