+++ src/main/java/org/influxdb/annotation/Column.java
@@ -18,22 +18,21 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.influxdb.pojomapper.annotation;
+package org.influxdb.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.util.concurrent.TimeUnit;
 
 /**
  * @author fmachado
  */
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Measurement {
+@Target(ElementType.FIELD)
+public @interface Column {
 
   String name();
 
-  TimeUnit timeUnit() default TimeUnit.MILLISECONDS;
+  boolean tag() default false;
 }
