+++ src/main/java/org/influxdb/impl/InfluxDBMapperException.java
@@ -18,7 +18,7 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.influxdb.pojomapper.exception;
+package org.influxdb.impl;
 
 /**
  * @author fmachado
